:recycle:  remove circular dependencies graphql package
