+++ packages/graphql/src/directive.ts
@@ -13,14 +13,12 @@ import type {
   CustomDirectiveOptions,
   Maybe,
   SchemaMap,
-  GraphQLDirective,
   CustomDirectiveMapItem,
 } from "@graphql-markdown/types";
 
 import { isEmpty } from "@graphql-markdown/utils";
 
 import { getDirective } from "./introspection";
-import { DirectiveLocation } from "graphql/language";
 
 /**
  * Wildcard `*` character for matching any directive name.
@@ -275,40 +273,4 @@ export const getConstDirectiveMap = (
     },
     {},
   );
-};
-
-/**
- * Checks if a directive is executable (related to operations).
- *
- * @param type - a GraphQL directive.
- *
- */
-export const executableDirectiveLocation = (
-  directive: GraphQLDirective,
-): boolean => {
-  return [
-    DirectiveLocation.QUERY,
-    DirectiveLocation.MUTATION,
-    DirectiveLocation.SUBSCRIPTION,
-    DirectiveLocation.FIELD,
-    DirectiveLocation.FRAGMENT_DEFINITION,
-    DirectiveLocation.FRAGMENT_SPREAD,
-    DirectiveLocation.INLINE_FRAGMENT,
-    DirectiveLocation.VARIABLE_DEFINITION,
-  ].some((item) => {
-    return directive.locations.includes(item);
-  });
-};
-
-/**
- * Checks if a directive is system (related to schema definition).
- *
- * @param type - a GraphQL directive.
- *
- */
-export const typeSystemDirectiveLocation = (
-  directive: GraphQLDirective,
-): boolean => {
-  // we assume that if not executable then it is a system directive
-  return !executableDirectiveLocation(directive);
 };
